+++ .eslintrc.json
@@ -16,6 +16,7 @@
     "ecmaFeatures": {
       "impliedStrict": true
     }
+//    "project": ["./tsconfig.json"]
   },
   "settings": {
     "noInlineConfig": true,
@@ -47,6 +48,7 @@
     "linebreak-style": "off",
     "node/no-unsupported-features/es-syntax": "off",
     "no-underscore-dangle": "off",
-    "import/prefer-default-export": "off"
+    "import/prefer-default-export": "off",
+    "dot-notation": "off"
   }
 }
