+++ nodemon.json
@@ -1,6 +1,6 @@
 {
   "ignore": ["node-modules", "test", "out"],
-  "watch": ["src/server.ts"],
-  "exec": "npm start",
+  "watch": ["./src/server.ts"],
+  "exec": "ts-node src/server.ts",
   "ext": "ts"
 }
