--- src/app.js
@@ -1,28 +1,0 @@
-const express = require('express');
-const swaggerUI = require('swagger-ui-express');
-const path = require('path');
-const YAML = require('yamljs');
-const userRouter = require('./resources/users/user.router');
-const boardsRouter = require('./resources/boards/boards.router');
-const tasksRouter = require('./resources/tasks/tasks.router');
-
-const app = express();
-const swaggerDocument = YAML.load(path.join(__dirname, '../doc/api.yaml'));
-
-app.use(express.json());
-
-app.use('/doc', swaggerUI.serve, swaggerUI.setup(swaggerDocument));
-
-app.use('/', (req, res, next) => {
-  if (req.originalUrl === '/') {
-    res.send('Service is running!');
-    return;
-  }
-  next();
-});
-
-app.use('/users', userRouter);
-app.use('/boards/:boardId/tasks', tasksRouter);
-app.use('/boards', boardsRouter);
-
-module.exports = app;
