+++ src/app.ts
@@ -1,0 +1,33 @@
+import express from 'express';
+import swaggerUI from 'swagger-ui-express';
+
+import path from 'path';
+import YAML from 'yamljs';
+
+import { fileURLToPath } from 'url';
+import { router as userRouter } from './resources/users/user.router';
+import { router as boardsRouter } from './resources/boards/boards.router';
+import { router as tasksRouter } from './resources/tasks/tasks.router';
+
+const dirname = path.dirname(fileURLToPath(import.meta.url));
+
+const app = express();
+const swaggerDocument = YAML.load(path.join(dirname, '../doc/api.yaml'));
+
+app.use(express.json());
+
+app.use('/doc', swaggerUI.serve, swaggerUI.setup(swaggerDocument));
+
+app.use('/', (req, res, next) => {
+  if (req.originalUrl === '/') {
+    res.send('Service is running!');
+    return;
+  }
+  next();
+});
+
+app.use('/users', userRouter);
+app.use('/boards/:boardId/tasks', tasksRouter);
+app.use('/boards', boardsRouter);
+
+export default app;
