+++ src/common/config.ts
@@ -10,5 +10,4 @@ dotenv.config({
 
 export const { PORT, NODE_ENV, MONGO_CONNECTION_STRING, JWT_SECRET_KEY } =
   process.env;
-
 export const AUTH_MODE = process.env['AUTH_MODE'] === 'true';
