+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "target": "es2020",
     "module": "es2020",
-    "allowJs": false,
+    "allowJs": true,
     "outDir": "./build",
     "rootDir": "./src",
     "downlevelIteration": false,
@@ -19,12 +19,15 @@
     "emitDecoratorMetadata": true,
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
-    "baseUrl": "./src"
+    "baseUrl": "./src",
+    "paths": {
+      "*.ts": ["*"]
+    }
   },
   "exclude": [
     "node_modules/"
   ],
   "include": [
-    "../src/**/*.ts"
+    "./src/**/*.ts"
   ]
 }
